Move Data mapping to DataRegistry
