+++ fe2-android/app/src/test/java/com/github/dedis/popstellar/utility/handler/ChirpHandlerTest.java
@@ -4,6 +4,7 @@ import static com.github.dedis.popstellar.utility.handler.MessageHandler.handleM
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import com.github.dedis.popstellar.di.DataRegistryModule;
 import com.github.dedis.popstellar.di.JsonModule;
 import com.github.dedis.popstellar.model.network.GenericMessage;
 import com.github.dedis.popstellar.model.network.answer.Result;
@@ -61,7 +62,7 @@ public class ChirpHandlerTest {
       new CreateLao(LAO_ID, LAO_NAME, CREATION_TIME, SENDER, new ArrayList<>());
   private static final AddChirp ADD_CHIRP = new AddChirp(TEXT, PARENT_ID, CREATION_TIME);
 
-  private static final Gson GSON = JsonModule.provideGson();
+  private static final Gson GSON = JsonModule.provideGson(DataRegistryModule.provideDataRegistry());
   private static final int REQUEST_ID = 42;
   private static final int RESPONSE_DELAY = 1000;
