+++ fe2-android/app/src/test/java/com/github/dedis/popstellar/utility/handler/ElectionHandlerTest.java
@@ -3,6 +3,7 @@ package com.github.dedis.popstellar.utility.handler;
 import static com.github.dedis.popstellar.utility.handler.ElectionHandler.electionSetupWitnessMessage;
 import static com.github.dedis.popstellar.utility.handler.MessageHandler.handleMessage;
 
+import com.github.dedis.popstellar.di.DataRegistryModule;
 import com.github.dedis.popstellar.di.JsonModule;
 import com.github.dedis.popstellar.model.network.GenericMessage;
 import com.github.dedis.popstellar.model.network.answer.Result;
@@ -60,7 +61,7 @@ public class ElectionHandlerTest extends TestCase {
   @Mock AndroidKeysetManager androidKeysetManager;
   @Mock PublicKeySign signer;
 
-  private static final Gson GSON = JsonModule.provideGson();
+  private static final Gson GSON = JsonModule.provideGson(DataRegistryModule.provideDataRegistry());
 
   private static final int REQUEST_ID = 42;
   private static final int RESPONSE_DELAY = 1000;
